moved functions out of notes file
