+++ bin/Cargo.toml
@@ -9,7 +9,6 @@ log = "0.4"
 clap = "2.33.1"
 simple_logger = "1.6.0"
 colored = "2"
-regex = "1"
 serde_json = "1.0.56"
 serde_derive = "1.0.114"
 serde = "1.0.114"
