+++ lib/src/lib.rs
@@ -15,10 +15,7 @@ extern crate mailparse;
 #[cfg(test)]
 extern crate mockall;
 extern crate colored;
-
-use note::LocalNote;
-
-use db::{DatabaseService, DBConnector};
+extern crate regex;
 
 #[macro_use]
 mod macros;
@@ -37,16 +34,16 @@ pub mod model;
 pub mod schema;
 pub mod builder;
 
+type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
 
-//use io::save_all_notes_to_file;
-/*pub fn fetch_all() {
-    let mut session = login();
-    let notes = apple_imap::fetch_notes(&mut session);
-    //save_all_notes_to_file(&notes);
-}*/
+use note::LocalNote;
+use db::{DatabaseService, DBConnector, SqliteDBConnection};
+use error::NoteError::NoteNotFound;
+use util::is_uuid;
+
 
 pub fn create_new_note<C>(db_connection: &dyn DatabaseService<C>, with_subject: String, folder: String)
-    -> Result<LocalNote,::error::NoteError> where C: 'static + DBConnector
+    -> std::result::Result<LocalNote,::error::NoteError> where C: 'static + DBConnector
 {
 
     let note = note!(
@@ -57,4 +54,40 @@ pub fn create_new_note<C>(db_connection: &dyn DatabaseService<C>, with_subject: 
     db_connection.insert_into_db(&note)
         .and_then(|_| Ok(note))
         .map_err(|e| ::error::NoteError::InsertionError(e.to_string()))
+}
+
+///Queries the database and tries to find a note with the provided search string
+/// Auto-Detects if the user provides the title or a uuid.
+///
+/// If multiple notes with the same title exist it returns the first one
+/// avaiable
+pub fn find_note(uuid_or_name: &String, db: &SqliteDBConnection) -> Result<LocalNote> {
+    match is_uuid(&uuid_or_name) {
+        true => {
+            match db.fetch_single_note(&uuid_or_name) {
+                Ok(Some(note)) => Ok(note),
+                Ok(None) => {
+                    eprintln!("Note does not exist");
+                    Err(NoteNotFound.into())
+                },
+                Err(e) => {
+                    eprint!("Error occured: {}", e.to_string());
+                    Err(e.into())
+                }
+            }
+        }
+        false => {
+            match db.fetch_single_note_with_name(&uuid_or_name) {
+                Ok(Some(note)) => Ok(note),
+                Ok(None) => {
+                    eprintln!("Note does not exist");
+                    Err(NoteNotFound.into())
+                },
+                Err(e) => {
+                    eprint!("Error occured: {}", e.to_string());
+                    Err(e.into())
+                }
+            }
+        }
+    }
 }
